Removed UpdateOnCreate.  Refactored AuditLogger
