+++ elide-integration-tests/src/test/java/example/TestCheckMappings.java
@@ -6,7 +6,6 @@
 package example;
 
 import com.yahoo.elide.security.checks.Check;
-import com.yahoo.elide.security.checks.prefab.Common;
 import com.yahoo.elide.security.checks.prefab.Role;
 
 import com.google.common.collect.ImmutableMap;
@@ -33,7 +32,6 @@ public class TestCheckMappings {
                     .put("child4Parent5", Child4Parent5Check.class)
                     .put("checkActsLikeFilter", AnotherFilterExpressionCheckObj.CheckActsLikeFilter.class)
                     .put("noRead", CreateButNoRead.NOREAD.class)
-                    .put("updateOnCreate", Common.UpdateOnCreate.class)
                     .put("checkLE", FilterExpressionCheckObj.CheckLE.class)
                     .put("checkRestrictUser", FilterExpressionCheckObj.CheckRestrictUser.class)
                     .put("specialValue", SpecialRead.SpecialValue.class)
